requirements.txt
aiogram>=3.24.0,<4.0
openai>=2.15.0
python-dotenv>=1.2.1
sqlalchemy>=2.0.46,<3.0
aiosqlite==0.22.1
python-dotenv==1.2.1